Update README.md
The rating column in the dataset shows -- as there is no rating values below 1.0
There are a few names of the restaurant with the same name in upper case and lower case.
The restaurant data contains cuisine in pairs like Chinese, and Indian. So it is necessary to provide unique cuisine details.
The id values for the restaurant are unique.
The cost column in the dataset contains values like ₹300,350, ₹8,000, ₹7,000, and ₹5,000 where the price is incorrect.
After placing the order, most of the customers have not given ratings to the restaurant.